Add ability to dynamically create FIX sessions
In order to support dynamic session creation in all Acceptor classes, an extra
refactor was done to create a `doAccept` function, like the `doConnect` version
in the Initiator classes.  Additionally, all `Log` and `Store` classes now use
the same `SessionSettings` instance as the Initiator / Acceptor.  In order to
make this possible, the Initiator / Acceptor use a non-const reference to
the settings object.  This seemed reasonable considering the `Application`,
`MessageStoreFactory`, and `LogFactory` are all passed in the same way.  This
could cause some clients to break if they are not holding onto the settings
object throughout the lifetime of the application.

If this change is too drastic, the other option is to keep separate settings
objects on each of the Store / Log / Acceptor / Initiator objects, but then
expose a function to create a session on each Store and Log.

Let me know if you have any questions or comments.